evol:add const and function voters #29

+++ src/Controller/UserController.php
@@ -2,22 +2,24 @@
 
 namespace App\Controller;
 
+use App\Entity\Task;
 use App\Entity\User;
+use App\Entity\UserType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use App\Form\UserType;
+
 /** 
-* @IsGranted("ROLE_ADMIN", message="No access! Get out!")
-*/
+ * @IsGranted("ROLE_ADMIN", message="No access! Get out!")
+ */
 class UserController extends AbstractController
 {
     /**
      * @Route("/users", name="user_list")
      */
-     
+
     public function listAction()
     {
         return $this->render('user/list.html.twig', ['users' => $this->getDoctrine()->getRepository('AppBundle:User')->findAll()]);
@@ -52,14 +54,15 @@ class UserController extends AbstractController
     /**
      * @Route("/users/{id}/edit", name="user_edit")
      */
-    public function editAction(User $user, Request $request)
+    public function editAction(Task $task, User $user, Request $request)
     {
+        $this->denyAccessUnlessGranted('task_edit', $task);
         $form = $this->createForm(UserType::class, $user);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
             $password = $this->get('security.password_encoder')->encodePassword($user, $user->getPassword());
             $user->setPassword($password);
 
@@ -73,11 +76,12 @@ class UserController extends AbstractController
         return $this->render('user/edit.html.twig', ['form' => $form->createView(), 'user' => $user]);
     }
 
-     /**
+    /**
      * @Route("/users/{id}/delete", name="user_delete")
      */
-    public function deleteUserAction(User $user)
+    public function deleteUserAction(Task $task,User $user)
     {
+        $this->denyAccessUnlessGranted('task_delete', $task);
         $em = $this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
